Change behaviour of tap-filter
Previously, tap-filter was redirecting output to stdout for successes
and stderr for test failures. It turns out this makes it hard to
distinguish between a tap-filter failure and a test failure.

This change more simply adds a -v switch. By default, only failures are
printed to stdout. With -v, all data will be printed.

This mimics the behaviour of `go test [-v]`, which seems sane.