tap-filter take two
The previous tap-filter was using a node library which meant debug lines
were silently discarded (because they are not part of the tap format).

The new tap-filter behaves the same way as the previous one (printing
only errors), except that in addition to printing the tap data it also
prints any output that was buffered in the failing test.

This allows us to see useful debug output on Travis, as well as
information about which test failing assertions are in.