add test/feature for bidirectional full syncs
This commit changes the way we handle requests in the integrationt tests.
We used to emit the request to the test harness before responding to the
SUT. This lead to sending an /admin/stats request to the SUT before sending
a response to the previous request to the sut.

Previously:
- receive join
- send /admin/stat request
- respond join

Now:
- receive join
- response join
- send /admin/stat request