Test throttling of bidirectional full sync  (#47)
As a safety check, both the ringpop-go and ringpop-node limit the number (defaults to 5) of concurrent bidirectional full syncs. 
A bidirectional full sync is triggered by a ping to a node without outstanding membership changes where the checksum in the ping doesn't match the node's checksum. This results in the node sending it's full membership. At the same time, the node will send a join to the source of the ping request to get it's full membership list.
This test pauses responding to the join request to make sure there are multiple joins happening at the same time. If the throttling is working, there should never be more than 5 joins in the queue.