Speed up integration tests; 10:18 to 3:12 minutes (#52)
Before waiting for the disseminator of the real node to drain we send it 30 pings (this will certainly drain it immediately). That means we no longer have to wait 6 seconds for the disseminator to drain every waitForEmptyPing is called.

It is certainly not the cleanest implementation but I really do not want to spent a lot of time on this.

The speedup in from 10:18 to 3:12 minutes. Thats a speedup of 3.2x.