Verify if source of changes is not triggering a bi-directional full-sync (#56)
During partition healing the changes send to target of the heal-operation are constructed in such a way that unmerge-able nodes in the memberships are marked as suspect so they will reincarnate. If the source of those changes is equal to the host-port of the node triggering the heal-operation, this could trigger a reverse full-sync. The reverse full-sync subsequently could result in a state where the two partitions start communicating before all the unmerge-able nodes are reincarnated. This would result in a lot of unnecessary gossiping.